[e-acsl] now using size_t instead of ulong

+++ src/plugins/e-acsl/share/e-acsl/memory_model/e_acsl_bittree.h
@@ -1,6 +1,8 @@
 #ifndef E_ACSL_BITTREE
 #define E_ACSL_BITTREE
 
-unsigned long mask(void*, void*);
+#include "stdlib.h"
+
+size_t mask(void*, void*);
 
 #endif
